+++ apps/web/src/widgets/place-review-summary/model/types.ts
@@ -15,26 +15,10 @@ export type PlaceReviewMetric = {
 };
 
 export type PlaceReviewScoreCounts = Record<ReviewEnvironmentScore, number>;
-export type UserReviewEnvironmentSummary = Omit<PlaceReviewMetric, 'count'>;
 
 export type PlaceReviewSummary = {
   totalCount: number;
   averageRating: number;
   ratingParticipantCount: number;
   items: PlaceReviewMetric[];
-};
-
-export type UserReviewInfo = {
-  memberKey: string;
-  nickname: string;
-  profileImageUrl: string;
-  rating: number;
-  createdAt: string;
-  environmentSummaries: UserReviewEnvironmentSummary[];
-  content?: string;
-  images?: string[];
-};
-
-export type PlaceReviewLists = PlaceReviewSummary & {
-  reviewLists: UserReviewInfo[];
 };
